Enhance motion sensor functionality and improve encounter log handling
- Added motion_sensor_remap_sample function to handle axis swapping and inversion for motion sensor data.
- Updated motion_sensor_fetch_sample to call the new remap function and ensure gyro data validity.
- Changed return type of find_slot_locked from ssize_t to int for consistency in encounter log management.
- Refactored encounter log functions to use int for slot variable instead of ssize_t.
- Introduced canvas_set_px_no_invalidate function to optimize pixel setting on canvas without invalidating the entire canvas.
- Updated draw_bitmap to utilize the new canvas pixel setting function.
- Updated submodule references for net-tools and zephyr.

+++ src/nearby/encounter_log.c
@@ -34,7 +34,7 @@ void encounter_log_init(void)
 	LOG_INF("Encounter log init (size=%d)", ENCOUNTER_LOG_SIZE);
 }
 
-static ssize_t find_slot_locked(uint32_t encounter_id)
+static int find_slot_locked(uint32_t encounter_id)
 {
 	size_t i;
 
@@ -44,7 +44,7 @@ static ssize_t find_slot_locked(uint32_t encounter_id)
 
 	for (i = 0U; i < ENCOUNTER_LOG_SIZE; i++) {
 		if (g_records[i].encounter_id == encounter_id) {
-			return (ssize_t)i;
+			return (int)i;
 		}
 	}
 
@@ -93,7 +93,7 @@ uint32_t encounter_log_begin(uint32_t ephemeral_id, uint8_t character_id, int64_
 
 void encounter_log_update(uint32_t encounter_id, int8_t rssi)
 {
-	ssize_t slot;
+	int slot;
 	struct encounter_record *rec;
 	int32_t sample;
 	int32_t avg;
@@ -131,7 +131,7 @@ void encounter_log_update(uint32_t encounter_id, int8_t rssi)
 
 void encounter_log_set_type(uint32_t encounter_id, enum kerfur_encounter_type type)
 {
-	ssize_t slot;
+	int slot;
 
 	k_mutex_lock(&g_log_mutex, K_FOREVER);
 	slot = find_slot_locked(encounter_id);
@@ -143,7 +143,7 @@ void encounter_log_set_type(uint32_t encounter_id, enum kerfur_encounter_type ty
 
 void encounter_log_mark_reaction(uint32_t encounter_id)
 {
-	ssize_t slot;
+	int slot;
 
 	k_mutex_lock(&g_log_mutex, K_FOREVER);
 	slot = find_slot_locked(encounter_id);
@@ -155,7 +155,7 @@ void encounter_log_mark_reaction(uint32_t encounter_id)
 
 void encounter_log_end(uint32_t encounter_id, int64_t now_ms)
 {
-	ssize_t slot;
+	int slot;
 
 	k_mutex_lock(&g_log_mutex, K_FOREVER);
 	slot = find_slot_locked(encounter_id);
@@ -191,7 +191,7 @@ size_t encounter_log_pending(struct encounter_record *out, size_t max)
 
 void encounter_log_mark_uploaded(uint32_t encounter_id)
 {
-	ssize_t slot;
+	int slot;
 
 	k_mutex_lock(&g_log_mutex, K_FOREVER);
 	slot = find_slot_locked(encounter_id);
